requirements.txt
pandas~=2.2.3
scikit-learn~=1.2.2
numpy~=1.26.0
requests~=2.32.3
tensorflow~=2.18.0
Flask~=3.0.3
torch~=2.5.1
PyYAML~=6.0.2
pillow~=11.0.0
torchvision~=0.20.1
matplotlib~=3.9.2
ultralytics~=8.3.38
opencv-python~=4.10.0.84
seaborn~=0.13.2
scipy~=1.14.1
tqdm~=4.67.1
psutil~=6.1.0
thop~=0.1.1-2209072238
Werkzeug~=3.1.2